% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{check_file_inputs}
\alias{check_file_inputs}
\alias{check_file_inputs.peaks}
\title{check_file_inputs
Validate values from config$input - called by summarizeMS}
\usage{
check_file_inputs(input, format)

check_file_inputs.peaks(input)
}
\arguments{
\item{input}{Character vector specifying location of input data}

\item{format}{Character value specifying format of input data}
}
\value{
Returns an updated character vector for input
}
\description{
check_file_inputs
Validate values from config$input - called by summarizeMS
}
