% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{generate_html_output}
\alias{generate_html_output}
\title{generate_html_output
Generate html summary files for MS data}
\usage{
generate_html_output(input, qmd)
}
\arguments{
\item{input}{Location of input files. `summary.html` and related files will also be stored in this location.}

\item{qmd}{Location of the qmd file to use. By default, `inst/quarto/summary.qmd` will be used.}
}
\description{
generate_html_output
Generate html summary files for MS data
}
\details{
Generates an html summary file using the RData file generated and saved in `process_file_inputs` and `qmd`
}
