% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{process_file_inputs}
\alias{process_file_inputs}
\alias{process_file_inputs.peaks}
\title{process_file_inputs
Process data - called by summarizeMS}
\usage{
process_file_inputs(input, format)

process_file_inputs.peaks(input)
}
\arguments{
\item{input}{Character vector specifying location of input data}

\item{format}{Character value specifying format of input data}
}
\value{
A list containing processed data
}
\description{
process_file_inputs
Process data - called by summarizeMS
}
