% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarizeMS}
\alias{summarizeMS}
\title{summarizeMS
Main function to run MSformatR}
\usage{
summarizeMS(input = NULL, format = NULL, qmd = NULL)
}
\arguments{
\item{input}{Character vector specifying the location of the data to be summarized. By default all MA output in the current working directory will be used. Use of unix-style path seperators, /, is required. R will check the system and replace them with the correct path separator if running on Windows (see Details for additional information).}

\item{format}{Character value specifying the source of the raw data. Default is 'peaks11' (see Details).}

\item{qmd}{Character value specifying the path to supporting quarto documents. Default value is `./summary.qmd` (see Details).}
}
\description{
summarizeMS
Main function to run MSformatR
}
\details{
When values are not supplied for function parameters, `summarizeMS` will look in the current working directory for a file called "config.yml" to supply these parameters. If no yaml config file is found, the default values specified here will be used instead.

The path(s) for `input` can be specified using one of three formats:

- An array of input directories (example yaml entry: `input: ["data/A", "data/B", "data/C"]`)
- A single path (example yaml entry: `input: "data/raw_results"`)
- A path to a directory containing multiple files to be summarized together. This should end in an `*`. (example yaml entry: `input: "data/*"`)

Supported formats:

- "peaks11" (default): html output from Peaks 11. Expected values for input are directory paths, each of which should contain multiple html files, csv files, and a directory of images.

If no qmd file is provided, MSformatR will look for supporting quarto documents in the current working directory. If the required supporting files are not found, it will use documents installed with this package.

Summary html pages will be generated by `summaryizeMS` as well as merged excel files when summarizing input from multiple samples.
}
